# backend/contracts/openapi/v1/provider.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "TAKEIN Provider API",
    "version": "1.0.0",
    "description": "Existing provider-owner and provider-branch API contract. Every operation requires JWT authentication plus the server-side provider account, branch, and menu permission checks identified by x-provider-permission."
  },
  "servers": [
    {
      "url": "https://api.takein.id/api",
      "description": "Production API ingress"
    }
  ],
  "paths": {
    "/provider/branch-profile": {
      "get": {
        "operationId": "showProviderBranchProfile",
        "summary": "Show the authenticated branch account profile",
        "description": "Returns the branch, provider, account, role, verification-document summaries, and profile completion for the branch_id carried by the authenticated provider JWT.",
        "x-authentication": "bearerAuth",
        "x-provider-permission": "profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated branch profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Only an authenticated branch account may access this profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateProviderBranchProfile",
        "summary": "Update the authenticated branch account profile",
        "description": "Updates only the branch carried by the authenticated JWT. Lifecycle, ownership, media-object, role, and account-login fields cannot be changed by this operation.",
        "x-authentication": "bearerAuth",
        "x-provider-permission": "profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Branch profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Branch scope or profile permission denied"
          },
          "422": {
            "description": "Validation failed"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchProfileUpdateRequest"
              }
            }
          }
        }
      }
    },
    "/provider/profile": {
      "get": {
        "operationId": "showProviderProfile",
        "summary": "Show the current provider profile",
        "x-authentication": "bearerAuth",
        "x-provider-permission": "profile",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Provider scope or permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateProviderProfile",
        "summary": "Update the current provider profile",
        "x-authentication": "bearerAuth",
        "x-provider-permission": "profile",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Provider scope or permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ProviderProfileUpdateRequest"
              }
            }
          }
        }
      }
    },
    "/provider/profile/documents/{document}": {
      "get": {
        "operationId": "downloadProviderDocument",
        "summary": "Download an owned private provider document",
        "x-authentication": "bearerAuth",
        "x-provider-permission": "owner",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Private document bytes",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Provider scope or permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "document",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ktp",
                "nib"
              ]
            }
          },
          {
            "name": "expires",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "signature",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/provider/profile/documents": {
      "post": {
        "operationId": "updateProviderDocuments",
        "summary": "Submit provider verification documents",
        "x-authentication": "bearerAuth",
        "x-provider-permission": "profile",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Provider scope or permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ProviderDocumentUpdateRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "JsonObject": {
        "type": "object",
        "additionalProperties": true
      },
      "ProviderProfileUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "email"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "username": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255
          },
          "phone_number": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 255
          },
          "image": {
            "type": [
              "string",
              "null"
            ],
            "format": "binary",
            "description": "Allowed media: JPEG, PNG, or WebP; maximum 2 MiB."
          }
        }
      },
...
    }
  }
}